movne_chatbot/requirements.txt
langchain==0.1.9
langchain-openai==0.0.8
openai==1.12.0
flask==3.0.2
pypdf==3.17.1
docx2txt==0.8
python-dotenv==1.0.0
firebase-admin==6.2.0
psutil==5.9.8
flask-cors==4.0.0
langchain-community==0.0.24
huggingface-hub==0.21.3 